requirements.txt
# Paper-Analysis 项目依赖包
# 安装方法: pip install -r requirements.txt

# 核心依赖 - PDF处理和Grobid客户端
grobid-client-python

# HTML/XML解析
beautifulsoup4

# 网络请求
requests

# 数据处理和分析
pandas

# 配置文件处理
PyYAML

# 豆包API SDK
volcengine-python-sdk[ark]

# arxiv 论文API
arxiv